Logistic Regression/requirements.txt
asgiref==3.2.7
cycler==0.10.0
Django==3.0.6
h5py==2.10.0
image==1.5.31
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scipy==1.4.1
six==1.15.0
sqlparse==0.3.1
